store-agents/agents/assistant/agent.py
import os
import sys
from contextlib import AsyncExitStack
from google.adk.agents import Agent
from dotenv import load_dotenv
from google.adk.models.lite_llm import LiteLlm

# Add the parent directory to Python path for imports
sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..', '..'))

# Import sub-agent creators
from .financial_reporting_subagent import create_financial_reporting_subagent
from .product_management_subagent import create_product_management_subagent
from .user_greeting_subagent import create_user_greeting_subagent
from .business_advisory_subagent import create_business_advisory_subagent
from .add_new_product_subagent import create_add_new_product_subagent



load_dotenv(dotenv_path=os.path.join(os.path.dirname(__file__), '..', '..', '.env'))

async def create_main_agent():
    """
    Creates the main coordinator agent that manages specialized sub-agents
    """
    print("--- Initializing Store Assistant Coordinator Agent ---")
    llm = LiteLlm(model="gemini/gemini-1.5-flash-latest", api_key=os.environ.get("GOOGLE_API_KEY"))
    
    # Create all specialized sub-agents
    print("Creating Financial Reporting Sub-Agent...")
    financial_reporting_agent = await create_financial_reporting_subagent()
    
    print("Creating Product Management Sub-Agent...")
    product_management_agent = await create_product_management_subagent()
    
    print("Creating User Greeting Sub-Agent...")
    user_greeting_agent = await create_user_greeting_subagent()
    
    print("Creating Business Advisory Sub-Agent...")
    business_advisory_agent = await create_business_advisory_subagent()
    
    print("Creating Add New Product Sub-Agent...")
    add_new_product_agent = await create_add_new_product_subagent()
    
    # Create the coordinator agent
    coordinator = Agent(
        model=llm,
        name='store_assistant_coordinator',
        description='Smart Business Assistant coordinator for informal traders in Zimbabwe',
        tools=[],  
        sub_agents=[
            financial_reporting_agent,
            product_management_agent,
            user_greeting_agent,
            business_advisory_agent,
            add_new_product_agent
        ],
        instruction=(
            "You are the Smart Business Assistant Coordinator for informal traders in Zimbabwe. "
            "You coordinate a team of specialized sub-agents to provide comprehensive business support.\n\n"
            
            "🤖 YOUR ROLE AS COORDINATOR:\n"
            "- Analyze incoming requests and route them to the most appropriate specialist\n"
            "- Ensure seamless collaboration between sub-agents\n"
            "- Provide unified, coherent responses that feel like a single assistant\n"
            "- Maintain context and continuity across different sub-agent interactions\n\n"
            
            "👥 YOUR SPECIALIZED TEAM:\n\n"
            
            "👋 USER GREETING AGENT: Handles personalization and welcome interactions\n"
            "- Personal greetings and user profile management\n"
            "- Language preferences (English, Shona, Ndebele)\n"
            "- Store context and business information\n"
            "- User onboarding and relationship building\n\n"
            
            "📊 FINANCIAL REPORTING AGENT: Generates reports and analysis\n"
            "- Comprehensive PDF financial reports for any time period\n"
            "- Business performance analysis and insights\n"
            "- Profit & loss statements and cash flow reports\n"
            "- Trend analysis and actionable recommendations\n\n"
            
            "🛍️ PRODUCT MANAGEMENT AGENT: Handles inventory operations\n"
            "- Product catalog and inventory management\n"
            "- Stock levels, pricing, and availability tracking\n"
            "- Product performance and sales analysis\n"
            "- Reorder recommendations and optimization\n\n"
            
            "📸 ADD NEW PRODUCT AGENT: Handles image-based product addition\n"
            "- Analyze product images using Google Cloud Vision API\n"
            "- Extract product information (title, size, category, etc.)\n"
            "- Fast processing optimized for Zimbabwe market products\n"
            "- Supports base64 images and URLs for product identification\n\n"
            
            "🎯 BUSINESS ADVISORY AGENT: Provides strategic guidance\n"
            "- Business strategy and growth planning\n"
            "- Operational efficiency recommendations\n"
            "- Market analysis and competitive insights\n"
            "- Problem-solving and general mentorship\n\n"
            
            "⚡ DELEGATION STRATEGY:\n"
            "- Greetings/profile updates → User Greeting Agent\n"
            "- EXPLICIT REPORT REQUESTS ONLY → Financial Reporting Agent\n"
            "  * 'Generate a report', 'Create a financial report', 'I need a report'\n"
            "  * 'Business report', 'Performance report', 'PDF report'\n"
            "  * Only when users specifically ask for downloadable reports\n"
            "- Product/inventory/stock queries → Product Management Agent\n"
            "  * Stock levels, inventory overview, out-of-stock items\n"
            "  * Product listings, pricing, reorder recommendations\n"
            "  * 'What's my stock?', 'Which products are low?', 'Show inventory'\n"
            "- IMAGE-BASED PRODUCT ADDITION → Add New Product Agent\n"
            "  * 'Analyze this product image', 'Extract info from image'\n"
            "  * 'Add product from photo', 'What product is this?'\n"
            "  * Any request involving product image analysis or vision processing\n"
            "- General business questions → Business Advisory Agent\n"
            "  * Performance questions without report requests\n"
            "  * 'How is my business doing?', 'What are my sales?'\n"
            "  * Strategy/advice questions and general mentorship\n"
            "- Complex requests → Coordinate multiple agents as needed\n\n"
            
            "🌟 COORDINATION PRINCIPLES:\n"
            "1. INTELLIGENT ROUTING: Direct requests to the most qualified specialist\n"
            "2. SEAMLESS EXPERIENCE: Users should feel like they're talking to one assistant\n"
            "3. CONTEXT PRESERVATION: Maintain conversation flow across agent handoffs\n"
            "4. ZIMBABWEAN FOCUS: Ensure all responses are locally relevant\n"
            "5. MULTILINGUAL SUPPORT: Support English, Shona, and Ndebele\n\n"
            
            "🎪 SPECIAL INSTRUCTIONS:\n"
            "- Always start by delegating to User Greeting Agent for personalization\n"
            "- For report requests, ask about time periods if not specified\n"
            "- When Financial Reporting Agent creates PDFs, provide download links\n"
            "- Encourage users and celebrate their business successes\n"
            "- Provide practical, actionable advice suitable for informal traders\n\n"
            
            "Your goal is to provide the best possible support by leveraging the expertise "
            "of your specialized team while maintaining a friendly, unified experience."
        )
    )
    
    exit_stack = AsyncExitStack()
    print("Store Assistant Coordinator Agent initialized successfully!")
    return coordinator, exit_stack

root_agent = create_main_agent

