store-agents/requirements.txt
google-generativeai
google-adk
google-cloud-vision
reportlab
firebase-admin
python-dotenv
litellm
requests
